% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hpml.R
\name{hpml.heatmap}
\alias{hpml.heatmap}
\title{Generate a heatmap from a dataframe}
\usage{
hpml.heatmap(df, pdf.filename, title)
}
\arguments{
\item{df}{A dataframe containing the data to be visualized. Must have a 
"label" column and partition columns.}

\item{pdf.filename}{The filename where the PDF will be saved.}

\item{title}{The title of the heatmap plot.}
}
\value{
A PDF file with the heatmap.
}
\description{
This function generates a heatmap from a dataframe containing partition and 
cluster information and saves it as a PDF.
}
\examples{
df <- data.frame(label = c("A", "B", "C"), partition1 = c(1, 2, 3))
hpml.heatmap(df, "heatmap.pdf", "My Heatmap")

}
