% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hpml.R
\name{hpml.save.plot}
\alias{hpml.save.plot}
\title{Save plot as PDF}
\usage{
hpml.save.plot(df, partition_col, filename, width, height, units, dpi)
}
\arguments{
\item{df}{Data frame containing the data with labels and partition information.}

\item{partition_col}{The name of the column in 'df' representing the partition 
or cluster to be visualized.}

\item{filename}{The name of the file where the plot will be saved (including path).}

\item{width}{The width of the output file.}

\item{height}{The height of the output file.}

\item{units}{The units for width and height (e.g., "in" for inches, "cm" for centimeters).}

\item{dpi}{The resolution of the output file in dots per inch.}
}
\value{
The plot is saved to the specified file, with no return value.
}
\description{
This function saves a plot generated by `plot.hpml` as a PDF file with the 
specified dimensions and resolution.
}
\examples{
df <- data.frame(label = c("A", "B", "C"), cluster = c(1, 2, 3))
hpml.save.plot(df, "cluster", "output.pdf", width = 8, height = 8, units = "in", dpi = 300)

}
